Change percent type on Float

+++ src/main/java/com/example/backendsynergyproject/dto/CategoryDto.java
@@ -2,5 +2,5 @@ package com.example.backendsynergyproject.dto;
 
 public record CategoryDto(String name,
                           Integer fullCost,
-                          Integer spendPercentOfBudgetCategory) {
+                          Float spendPercentOfBudgetCategory) {
 }
