+++ src/main/java/com/example/backendsynergyproject/models/Category.java
@@ -28,7 +28,7 @@ public class Category {
     private Integer fullCost;
 
     @NotNull
-    private Integer spendPercentOfBudgetCategory;
+    private Float spendPercentOfBudgetCategory;
 
 
     @OneToMany(cascade = CascadeType.ALL, orphanRemoval = true)
